tests/wpt/tests/compute-pressure/permissions-policy/compute-pressure-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that compute-pressure is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissions-policy-features">
<link rel="help" href="https://w3c.github.io/compute-pressure/#policy-control">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('compute-pressure', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise compute-pressure.');
</script>
